chore: add e2e test for single payment export
AB#37532

Besides adding the test, I've also split `PaymentsPage` into `PaymentsPage` and `PaymentPage`, because a lot of the things in `PaymentsPage` were actually the payment details page, not the payments overview page.
