+++ e2e/portal/pages/PaymentsPage.ts
@@ -14,16 +14,7 @@ class PaymentsPage extends BasePage {
   readonly paymentTitle: Locator;
   readonly paymentSummaryMetrics: Locator;
   readonly paymentSummaryWithInstructions: Locator;
-  readonly exportFspPaymentListButton: Locator;
   readonly exportButton: Locator;
-  readonly importReconcilationDataButton: Locator;
-  readonly chooseFileButton: Locator;
-  readonly importFileButton: Locator;
-  readonly proceedButton: Locator;
-  readonly viewPaymentTitle: Locator;
-  readonly paymentAmount: Locator;
-  readonly retryFailedTransfersButton: Locator;
-  readonly popupRetryTransferButton: Locator;
 
   constructor(page: Page) {
     super(page);
@@ -46,31 +37,8 @@ class PaymentsPage extends BasePage {
     this.paymentSummaryWithInstructions = this.page.getByTestId(
       'create-payment-excel-fsp-instructions',
     );
-    this.exportFspPaymentListButton = this.page.getByLabel(
-      'Export FSP payment list',
-    );
     this.exportButton = this.page.getByRole('button', {
       name: 'Export',
-    });
-    this.importReconcilationDataButton = this.page.getByRole('button', {
-      name: 'Import reconciliation data',
-    });
-    this.chooseFileButton = this.page.getByRole('button', {
-      name: 'Choose file',
-    });
-    this.importFileButton = this.page.getByRole('button', {
-      name: 'Import file',
-    });
-    this.proceedButton = this.page.getByRole('button', { name: 'Proceed' });
-    this.viewPaymentTitle = this.page.getByRole('heading', {
-      name: 'Payment',
-    });
-    this.paymentAmount = this.page.getByTestId('metric-tile-component');
-    this.retryFailedTransfersButton = this.page.getByRole('button', {
-      name: 'Retry failed transfer(s)',
-    });
-    this.popupRetryTransferButton = this.page.getByRole('button', {
-      name: 'Retry transfers',
     });
   }
 
@@ -110,33 +78,12 @@ class PaymentsPage extends BasePage {
     await this.startPaymentButton.click();
   }
 
-  async waitForPaymentToComplete() {
-    const inProgressChip = this.page
-      .locator('app-colored-chip')
-      .getByLabel('In progress');
-
-    await inProgressChip.waitFor({ state: 'hidden' });
-  }
-
-  async validateInProgressChipIsPresent() {
-    const inProgressChip = this.page
-      .locator('app-colored-chip')
-      .getByLabel('In progress');
-
-    await expect(inProgressChip).toBeVisible();
-  }
-
   async validateInProgressBannerIsPresent() {
     const inProgressBanner = this.page.getByTestId(
       'payment-in-progress-banner',
     );
 
     await expect(inProgressBanner).toBeVisible();
-  }
-
-  async validatePaymentsDetailsPageByDate(date: string) {
-    const viewPaymentTitle = await this.viewPaymentTitle.textContent();
-    expect(viewPaymentTitle).toContain(date);
   }
 
   async validateNumericValue(
@@ -238,75 +185,6 @@ class PaymentsPage extends BasePage {
   async validateExportMessage({ message }: { message: string }) {
     const exportMessage = this.page.getByText(message);
     await expect(exportMessage).toBeVisible();
-  }
-
-  async validateGraphStatus({
-    pending,
-    successful,
-    failed,
-  }: {
-    pending: number;
-    successful: number;
-    failed: number;
-  }) {
-    await this.page.waitForTimeout(1000); // Wait for the graph to be updated after the loader is hidden
-    const graph = await this.page.locator('canvas').getAttribute('aria-label');
-    if (graph) {
-      const graphText = graph
-        .replace('Payment status chart.', '')
-        .replace(/\s+/g, ' ')
-        .trim();
-
-      expect(graphText).toContain(
-        `Pending: ${pending}, Successful: ${successful}, Failed: ${failed}`,
-      );
-    } else {
-      throw new Error('Graph attribute is null');
-    }
-  }
-
-  async validateTransferAmounts({ amount }: { amount: number }) {
-    await this.page.waitForTimeout(1000); // Wait for the graph to be updated after the loader is hidden
-
-    const paymentAmount = this.paymentAmount.getByText('€');
-    const paymentAmountText = await paymentAmount.textContent();
-    const normalizedPaymentAmountText = paymentAmountText?.replace(/,/g, '');
-
-    expect(normalizedPaymentAmountText).toContain(amount.toString());
-  }
-
-  async validateRetryFailedTransfersButtonToBeVisible() {
-    await expect(this.retryFailedTransfersButton).toBeVisible();
-  }
-
-  async validateRetryFailedTransfersButtonToBeHidden() {
-    await expect(this.retryFailedTransfersButton).toBeHidden();
-  }
-
-  async retryFailedTransfers() {
-    await this.table.filterColumnByDropDownSelection({
-      columnName: 'Transfer status',
-      selection: 'Failed',
-    });
-
-    await this.table.selectAll();
-
-    await this.retryFailedTransfersButton.click();
-
-    await this.popupRetryTransferButton.click();
-  }
-
-  async importReconciliationData(filePath: string) {
-    await this.importReconcilationDataButton.click();
-
-    await this.chooseAndUploadFile(filePath);
-
-    await this.importFileButton.click();
-  }
-
-  async exportFspPaymentList() {
-    await this.exportButton.click();
-    await this.exportFspPaymentListButton.click();
   }
 
   async isPaymentPageEmpty(): Promise<boolean> {
