+++ e2e/portal/tests/DoPayment/DoSuccessfulPaymentWithCbe.spec.ts
@@ -16,6 +16,7 @@ import {
 } from '@121-service/test/registrations/pagination/pagination-data';
 
 import LoginPage from '@121-e2e/portal/pages/LoginPage';
+import PaymentPage from '@121-e2e/portal/pages/PaymentPage';
 import PaymentsPage from '@121-e2e/portal/pages/PaymentsPage';
 
 let accessToken: string;
@@ -32,6 +33,7 @@ test.beforeEach(async ({ page }) => {
 });
 
 test('[36081] Do successful payment for Cbe fsp', async ({ page }) => {
+  const paymentPage = new PaymentPage(page);
   const paymentsPage = new PaymentsPage(page);
   const projectTitle = CbeProgram.titlePortal.en;
   const numberOfPas = registrationsCbe.length;
@@ -63,13 +65,13 @@ test('[36081] Do successful payment for Cbe fsp', async ({ page }) => {
       url.pathname.startsWith(`/en-GB/project/${programIdCbe}/payments/1`),
     );
     // Assert payment overview page by payment date/ title
-    await paymentsPage.validatePaymentsDetailsPageByDate(lastPaymentDate);
+    await paymentPage.validatePaymentsDetailsPageByDate(lastPaymentDate);
   });
 
   await test.step('Validate payment card', async () => {
-    await paymentsPage.validateToastMessage('Payment created.');
-    await paymentsPage.navigateToProgramPage('Payments');
-    await paymentsPage.waitForPaymentToComplete();
+    await paymentPage.validateToastMessage('Payment created.');
+    await paymentPage.waitForPaymentToComplete();
+    await paymentPage.navigateToProgramPage('Payments');
     await paymentsPage.validatePaymentCard({
       date: lastPaymentDate,
       paymentAmount: defaultMaxTransferValue,
