+++ e2e/portal/tests/DoPayment/DoSuccessfulPaymentWithSafaricom.spec.ts
@@ -14,6 +14,7 @@ import {
 } from '@121-service/test/registrations/pagination/pagination-data';
 
 import LoginPage from '@121-e2e/portal/pages/LoginPage';
+import PaymentPage from '@121-e2e/portal/pages/PaymentPage';
 import PaymentsPage from '@121-e2e/portal/pages/PaymentsPage';
 
 test.beforeEach(async ({ page }) => {
@@ -32,6 +33,7 @@ test.beforeEach(async ({ page }) => {
 });
 
 test('[36009] Do successful payment for Safaricom fsp', async ({ page }) => {
+  const paymentPage = new PaymentPage(page);
   const paymentsPage = new PaymentsPage(page);
   const projectTitle = KRCSProgram.titlePortal.en;
   const numberOfPas = registrationsSafaricom.length;
@@ -60,13 +62,13 @@ test('[36009] Do successful payment for Safaricom fsp', async ({ page }) => {
       ),
     );
     // Assert payment overview page by payment date/ title
-    await paymentsPage.validatePaymentsDetailsPageByDate(lastPaymentDate);
+    await paymentPage.validatePaymentsDetailsPageByDate(lastPaymentDate);
   });
 
   await test.step('Validate payment card', async () => {
-    await paymentsPage.validateToastMessageAndClose('Payment created.');
-    await paymentsPage.navigateToProgramPage('Payments');
-    await paymentsPage.waitForPaymentToComplete();
+    await paymentPage.validateToastMessageAndClose('Payment created.');
+    await paymentPage.waitForPaymentToComplete();
+    await paymentPage.navigateToProgramPage('Payments');
     await paymentsPage.validatePaymentCard({
       date: lastPaymentDate,
       paymentAmount: defaultMaxTransferValue,
