+++ e2e/portal/tests/ViewPayment/ExportPaymentReport.spec.ts
@@ -1,5 +1,4 @@
 import { test } from '@playwright/test';
-import { format } from 'date-fns';
 
 import { SeedScript } from '@121-service/src/scripts/enum/seed-script.enum';
 import NLRCProgram from '@121-service/src/seed-data/program/program-nlrc-ocw.json';
@@ -7,13 +6,14 @@ import { seedIncludedRegistrations } from '@121-service/test/helpers/registratio
 import {
   getAccessToken,
   resetDB,
-  resetDuplicateRegistrations,
 } from '@121-service/test/helpers/utility.helper';
 import {
   programIdOCW,
   registrationOCW1,
+  registrationOCW6Fail,
 } from '@121-service/test/registrations/pagination/pagination-data';
 
+import ExportData from '@121-e2e/portal/components/ExportData';
 import LoginPage from '@121-e2e/portal/pages/LoginPage';
 import PaymentPage from '@121-e2e/portal/pages/PaymentPage';
 import PaymentsPage from '@121-e2e/portal/pages/PaymentsPage';
@@ -22,11 +22,10 @@ test.beforeEach(async ({ page }) => {
   await resetDB(SeedScript.nlrcMultiple, __filename);
   const accessToken = await getAccessToken();
   await seedIncludedRegistrations(
-    [registrationOCW1],
+    [registrationOCW1, registrationOCW6Fail],
     programIdOCW,
     accessToken,
   );
-  await resetDuplicateRegistrations(8);
 
   // Login
   const loginPage = new LoginPage(page);
@@ -34,13 +33,14 @@ test.beforeEach(async ({ page }) => {
   await loginPage.login();
 });
 
-test('[32296] Show in progress banner and chip when payment is in progress', async ({
+test('[35621] Export Payment Report should contain the right data', async ({
   page,
 }) => {
   const paymentPage = new PaymentPage(page);
   const paymentsPage = new PaymentsPage(page);
+  const exportDataComponent = new ExportData(page);
+
   const projectTitle = NLRCProgram.titlePortal.en;
-  const lastPaymentDate = `${format(new Date(), 'dd/MM/yyyy')}`;
 
   await test.step('Navigate to Program payments', async () => {
     await paymentsPage.selectProgram(projectTitle);
@@ -53,19 +53,19 @@ test('[32296] Show in progress banner and chip when payment is in progress', asy
     await paymentsPage.startPayment();
     // Assert redirection to payment overview page
     await page.waitForURL((url) =>
-      url.pathname.startsWith('/en-GB/project/3/payments/1'),
+      url.pathname.startsWith(`/en-GB/project/${programIdOCW}/payments/1`),
     );
-    // Assert payment overview page by payment date/ title
-    await paymentPage.validatePaymentsDetailsPageByDate(lastPaymentDate);
+    await paymentPage.waitForPaymentToComplete();
   });
 
-  await test.step('Validate payemnt in progress in Payment overview', async () => {
-    await paymentPage.validateToastMessage('Payment created.');
-    await paymentPage.validateInProgressChipIsPresent();
-  });
+  await test.step('Export Payment Report', async () => {
+    await paymentPage.selectPaymentExportOption({
+      option: 'Payment report',
+    });
 
-  await test.step('Validate payemnt in progress in Payments page', async () => {
-    await paymentPage.navigateToProgramPage('Payments');
-    await paymentsPage.validateInProgressBannerIsPresent();
+    await exportDataComponent.exportAndAssertData({
+      exactRowCount: 2,
+      excludedColumns: ['created', 'updated'],
+    });
   });
 });
