+++ e2e/portal/tests/ViewPayment/GraphDisplaysCorrectPaymentAmount.spec.ts
@@ -14,6 +14,7 @@ import {
 } from '@121-service/test/registrations/pagination/pagination-data';
 
 import LoginPage from '@121-e2e/portal/pages/LoginPage';
+import PaymentPage from '@121-e2e/portal/pages/PaymentPage';
 import PaymentsPage from '@121-e2e/portal/pages/PaymentsPage';
 
 test.beforeEach(async ({ page }) => {
@@ -30,6 +31,7 @@ test.beforeEach(async ({ page }) => {
 test('[32299] Table should reflect the actual transfer amounts sent to the PAs in this payment', async ({
   page,
 }) => {
+  const paymentPage = new PaymentPage(page);
   const paymentsPage = new PaymentsPage(page);
   const projectTitle = NLRCProgram.titlePortal.en;
   const lastPaymentDate = `${format(new Date(), 'dd/MM/yyyy')}`;
@@ -53,14 +55,14 @@ test('[32299] Table should reflect the actual transfer amounts sent to the PAs i
       url.pathname.startsWith(`/en-GB/project/${programIdOCW}/payments/1`),
     );
     // Assert payment overview page by payment date/ title
-    await paymentsPage.validatePaymentsDetailsPageByDate(lastPaymentDate);
+    await paymentPage.validatePaymentsDetailsPageByDate(lastPaymentDate);
   });
 
   await test.step('Validate transfer amount after "payment in progress" chip disappears in Payment overview', async () => {
-    await paymentsPage.validateToastMessage('Payment created.');
-    await paymentsPage.waitForPaymentToComplete();
+    await paymentPage.validateToastMessage('Payment created.');
+    await paymentPage.waitForPaymentToComplete();
 
-    await paymentsPage.validateTransferAmounts({
+    await paymentPage.validateTransferAmounts({
       amount: defaultMaxTransferValue,
     });
   });
