+++ e2e/portal/tests/__screenshots__/ViewPayment/ExportPaymentReport.spec.ts/exported-xlsx.csv
@@ -1,0 +1,2 @@
+referenceId,id,registrationId,status,payment,registrationStatus,phoneNumber,paymentAmountMultiplier,amount,errorMessage,fsp,fullName,whatsappPhoneNumber,addressStreet,addressHouseNumber,addressHouseNumberAddition,addressPostalCode,addressCity
+54e6286jZ4mhkTtgjM,2,4,error,1,included,14155235555,3,75,CREATE CUSTOMER ERROR: NOT_FOUND: We mocked that creating customer failed Field: mock field,Intersolve-visa,Test mock-fail-create-customer,14155235555,AnotherStreet,4,C,4567DE,AnotherCity
